XRBK: add `self::length` syntax
This adds syntax to `Source` `Arg`s such that message lengths (in `Request`s and `Reply`s) may be used in `Source`s.

This is useful especially when the length of a list is not encoded separately from the length of the message - the context attribute of that list may need to use the length of the message in order to deserialize the list.

Example:
```rust
#[context(self::length => ((*length as usize) * 4) - 8)]
```